requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126

# Fresh Windows inference environment target:
# - Python 3.10
# - CUDA Toolkit 12.8 or 12.9
# - TensorRT 10.15.1.29
# - PyTorch 2.6.0+cu126 for the main JoyVASA models
# - ONNX Runtime 1.21.0 for the cropper
# - xFormers 0.0.29.post3
# In this repo, ONNX Runtime is used only for the cropper with CUDA EP, while
# the main models use native TensorRT runners.
# This modern stack aligns PyTorch and ONNX Runtime on CUDA 12.x + cuDNN 9,
# which is the required direction if the cropper and the main model stack must
# share one Python process on Windows.

accelerate==0.28.0
audio-separator==0.17.2
av==12.1.0
bitsandbytes==0.43.1
decord==0.6.0
diffusers==0.27.2
einops==0.8.0
insightface==0.7.3
librosa==0.10.2.post1
mediapipe==0.10.14
mlflow==2.13.1
moviepy==1.0.3
numpy==1.26.4
omegaconf==2.3.0
onnx2torch==1.5.14
onnx==1.16.1
onnxruntime-gpu[cuda,cudnn]==1.21.0
tensorrt-cu12==10.15.1.29
opencv-python==4.10.0.84
pillow==10.3.0
pyyaml==6.0.1
setuptools==70.0.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.66.4
transformers==4.39.2
xformers==0.0.29.post3
isort==5.13.2
pylint==3.2.2
pre-commit==3.7.1
scipy==1.13.1
imageio==2.34.2
lmdb==1.4.1
rich==13.7.1
ffmpeg-python==0.2.0
scikit-image==0.24.0
albumentations==1.4.10
matplotlib==3.9.0
imageio-ffmpeg==0.5.1
tyro==0.8.5
gradio==5.1.0
pykalman==0.9.7
tensorboardX==2.6.2.2
